+++ tests/run-region.sh
@@ -6,7 +6,7 @@ LAZ=/media/tony/storpart/fpctrunklaztrunk/lazarus
 FPC=/media/tony/storpart/fpctrunklaztrunk/fpc/bin/x86_64-linux/fpc
 [ -x "$FPC" ] || FPC=fpc
 WS=${LCL_WS:-gtk3}
-"$FPC" -B -Mobjfpc -Sh -Cirot -O1 -FE/tmp -FU/tmp -Fu. \
+"$FPC" -B -Mobjfpc -Sh -Cirot -O1 -FE/tmp -FU/tmp -Fu. -Fuwebp \
   -Fu"$LAZ/lcl/units/x86_64-linux" \
   -Fu"$LAZ/lcl/units/x86_64-linux/$WS" \
   -Fu"$LAZ/components/lazutils/lib/x86_64-linux" \